backend/requirements.txt
langgraph
langchain-openai
python-dotenv
RestrictedPython
grandalf